h3: add fin flag to HTTP/3 Events
Previously, when `quiche::h3::Connection::poll()`
returned a `Headers` or `Data` event, there was
no way to tell if that event simultaneously closed
the indicated stream. Instead, `poll()` would need
to be called until the `Finished` event occurred.

With this change a boolean `fin` flag is returned
with the Event, which allows for some more
optimised work flows.